+++ locales/de.json
@@ -1,0 +1,22 @@
+{
+  "terms": {
+    "title": "AGB",
+    "content": "Dies ist die deutsche AGB-Seite."
+  },
+  "headers": {
+    "how-it-works": "Wie es funktioniert",
+    "categories": "Kategorien",
+    "trust-safety": "Vertrauen und Sicherheit",
+    "jobs": "Jobs",
+    "become-helper": "Werden Sie Helfer",
+    "post-mini-job": "Minijob veröffentlichen",
+    "view-all-notifications": "Alle Benachrichtigungen anzeigen",
+    "my-profile": "Mein Profil",
+    "my-jobs": "Meine Jobs",
+    "saved-jobs": "Gespeicherte Jobs",
+    "applied-jobs": "Angewandte Jobs",
+    "logout": "Abmelden",
+    "notifications": "Benachrichtigungen",
+    "close": "Schließen"
+  }
+}
