--- messages/de.json
@@ -1,6 +1,0 @@
-{
-  "terms": {
-    "title": "AGB",
-    "content": "Dies ist die deutsche AGB-Seite."
-  }
-}
