--- messages/en.json
@@ -1,6 +1,0 @@
-{
-  "terms": {
-    "title": "Terms & Conditions",
-    "content": "This is the English terms page."
-  }
-}
